+++ src/Counter.php
@@ -20,6 +20,26 @@ use Esi\SimpleCounter\Interface\CounterInterface;
  *
  * @see \Esi\SimpleCounter\Tests\CounterTest
  */
+
+/**
+ * PHPStan type for each options array definition, for each adapter.
+ *
+ * Placed here so that it can be imported via phpstan-import-type into
+ * where it is needed.
+ *
+ * @phpstan-type BaseAdapterOptions = array{
+ *     imageDir?: string,
+ *     imageExt?: string,
+ *     uniqueOnly?: bool,
+ *     asImage?: bool
+ * }|array{}
+ *
+ * @phpstan-type FlatfileOptions = array{
+ *     logDir?: string,
+ *     countFile?: string,
+ *     ipFile?: string
+ * }
+ */
 readonly class Counter
 {
     /**
@@ -61,5 +81,13 @@ readonly class Counter
     public function fetchCurrentIpList(): array
     {
         return $this->adapter->fetchCurrentIpList();
+    }
+
+    /**
+     * Returns the given option, if it exists.
+     */
+    public function getOption(string $option): string | bool | null
+    {
+        return $this->adapter->getOption($option);
     }
 }