+++ tests/bootstrap.php
@@ -10,6 +10,10 @@ declare(strict_types=1);
  * For the full copyright and license information, please view
  * the LICENSE.md file that was distributed with this source code.
  */
+
+/**
+ * @internal
+ */
 \error_reporting(\E_ALL);
 
 require_once \dirname(__DIR__) . '/vendor/autoload.php';